tdf#153808 Fix numbering preview in dark mode
This patch fixes the color used for numbers and bullets in the preview
of the Bullets and Numbering dialog in Impress. Currently the
implementation causes the bullets and numbers to be rendered in white
over a white background when using dark mode.

The fix uses DOCCOLOR and FONTCOLOR to create the preview, instead of
FieldColor and FieldTextColor.

Change-Id: Icf41c5be59c67a2d68e3c520744276492e9baa59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147664
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
(cherry picked from commit 3a4b32c9e1dfc56d9b00e3720834d21c9095f629)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149435
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>