sw floattable: allow extra space on top of child content in SwFlyFrame::Grow_()
Once a floating table with a single cell wanted to span over two pages
("split row"), SwTextFrame::SplitFrame() was called, but then the table
split failed because everything moved to the next page.

The reason for this was that we tried to grow the text frame (and all of
its parents: cell, row, table, fly), but a non-test growing a fly
normally just works by recalculating its size based on the content,
ignoring what growth is wanted.

This is fine in the non-split fly case, but we do want to increase the
size of a split fly to control what content remains in the current fly
frame and what goes to a follow fly frame.

Fix this by extending SwFlyFrame::Grow_() so it grows more than just its
own content in the split fly case if the request wanted that.  Now a
single row can split, but the position of the follow fly frame is still
bad.

(cherry picked from commit 7677e16217349a0a0e94edb6a90b00089432c6ce)

Change-Id: I1c6aa61996dab058c291331a795627d3fae1bca8