sw floattable: grow the print area as well in SwFlyFrame::Grow_()
The bugdoc should be of 3 pages, but the last for of the floating table
went to the 3rd page.

The reason for this was that SwTabFrame::MakeAll() calculated
nDistanceToUpperPrtBottom, then determined that the upper has to grow,
and called GetUpper()->Grow(LONG_MAX, /*bTest=*/true).

Now this growth was less than what would be possible, because a previous
invocation of SwFlyFrame::Grow_() increased the frame area (with
margins) but not the print area (with margins). This meant that the
frame can't grow too much because its frame height is already 7044 twips
but its print height is only 548. This combination of "not enough inner
space" and "can't grow too much, either" resulted in moving the last row
to an unwanted 3rd page.

Fix the problem by not only increasing the frame height but also the
print height in SwFlyFrame::Grow_(), this keeps the last row on the 2nd
page, like Word does.

(cherry picked from commit 5d5dca66e17c90e20197d0d76113254b13ff0bb7)

Change-Id: I7850b54b030c42034e0a117d52d2148fa6ba2f83