sw floattable: handle Word 2010 legacy mode in SwFlyFrame
The tdf#61594 bugdoc has 3 non-header rows on the first page, while
Writer splits the 3rd row, so the follow row goes to the 2nd page.

It turns out that the 3rd row fits the first page in Word because >=
2013 matches Writer, but <= 2010 used to behave differently, see
<https://answers.microsoft.com/en-us/msoffice/forum/all/floating-tables-in-word-2013-will-not-extend/576962cc-392d-43e5-b603-654406e1969f>.
(The spec doesn't seem to document this, sadly.)

Fix the problem by using one of the existing compat flags that we only
set in the Word <= 2010 case (triggered by <w:compatSetting
w:name="compatibilityMode" w:val="14"> or lower values), that scenario
wants to allow the fly frame to extend till the bottom of the page, not
only till the bottom of the body frame print area.

The bugdoc now doesn't split that 3rd row, but the layout of the 4th row
is still bad with this.

(cherry picked from commit da2707a83f13cba98b22aba1ca6568dbbc4c5fd8)

Change-Id: I547abb19df250e5fbe4d8e123593a645cdbdc6b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148835
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>