tdf#135581 sw UI: keep image in place when changing anchor
Changing the anchor type should not move an image around.

The only time we have a position to fix AFAICS is when
the anchor changes and the RelationOrient is no longer appropriate.
In that case we always get a new p*OrientItem.
If GetPos changed, then that means the user (or earlier algorithm)
already adjusted for the change. If there is no difference in GetPos
even though the relation changed, then that is the only time we (likely)
need to recalculate.

So the typical example requiring an adjustment would be a change
from "to paragraph" at text position 0 to a "to page" anchor.
In this case "page" has no relationship to "text margins"
and thus needs to base the absolute position 0 on page instead of text,
meaning that the margins would need to be added to the absolute position.

This code is basically unchanged since initial import.
I have no idea why it could possibly be good to change
the position when there is no new information given
(no pHoriOrientItem/pVertOrientItem).

[I checked and "from top" and "from bottom"
 are both == NONE, as well of course as "from left".]

Change-Id: Ib89ad9ba26c8826ed43f6d4505a82502b9ad8af9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149012
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149634
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>