Partial fix: iOS missing text direction listbox in Format > Page Style > Page
Add the static libraries that are required by the LibreOffice's
CTL and CJK text layout features to the iOS app add link list.

libnumbertext.a is one of the missing static libraries that are required by the LibreOffice's CTL and CJK text layout features.

The other missing static library (which was in the ios-all-static-libs list but was getting stripped by Xcode) is added back by listing the "lingucomponent_NumberText_get_implementation" C function.

Without these static library changes, enabling CTL or CJK will cause a crash as the "lingucomponent_NumberText_get_implementation" symbol won't be in the binary and that function needs libnumbertext.a.

Reference-to: https://github.com/CollaboraOnline/online/issues/6050
Change-Id: I21fddbfd29a1d326b509840127bd136c327cd3d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149240
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Patrick Luby <plubius@neooffice.org>
Reviewed-by: Andras Timar <andras.timar@collabora.com>