sw floattable: fix inner floating table inside normal outer table from DOCX
The problem was that CppunitTest_sw_ooxmlexport9's testTdf79329 has a
normal outer table and a floating inner table; the inner table was
already not floating (so that's not a new problem), but
SW_FORCE_FLY_SPLIT=1 even failed the text-to-table conversion for the
outer table, so we just had 1 table, not 2 tables.

The problem seems to be that the start/end positions for the outer table
cell get invalidated by the inner table-to-frame conversion, so the
outer table conversion will fail as well.

Fix the problem by limiting the table-to-frame conversion for toplevel
tables: this avoids the failing text-to-table conversion in the
SW_FORCE_FLY_SPLIT=1 case.

At some stage I should revisit this, since the DOC import has working
floating tables for the inner case, for now just make sure that the DOCX
import result has two tables, as before.

(cherry picked from commit a1b935ca1bb6d48241e73e7206a367fe2b51f948)

Change-Id: I39aa00e46c12a32117c334cb97e1cc0270b77651
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149374
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>