sw CopyBookmarks: ensure that Move copies can rename
The problem was that a reference to the bookmark name
was failing - since the move was adding a "Copy 1" at the end.

That rename line is very deceptive because AFAICS it fails every time.
I'm going to look at removing it in a followup commit.

Two other unit tests also found with make sw.check
-tdf149089.docx
-tdf149507.docx
but they also didn't have references to the bookmark to visually notice.

make CppunitTest_sw_ooxmlexport10 CPPUNIT_TEST_NAME=testTdf92157

Change-Id: Idd695ec4a89057b28a68bc051b73f17fd4c3de56
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149574
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149641
Tested-by: Justin Luth <jluth@mail.com>