tdf#154882: svx: don't overwrite the first cell's style
i.e. new Draw document, change the char formatting in the
default drawing style (e.g. font size), insert a table and
fill its cells with text. Save and reload. After reload the
first cell inherits the formatting of the default DS,
instead of its assigned cell style. But dragging the table
around shows the correct formatting in the dragging overlay,
and it's also correct when copy & pasting the table.

This is caused by the order of things at import, as a table
is created initially with a single cell at least, and it
gets its cell style assignment too early. There is also
similar code to what is changed here in SdrTextObj::
BegTextEdit, and it's rightfully overridden by SdrTableObj.

This appears to be a regression from 4-0, but I'm not sure
which commit and what exactly changed there.

Change-Id: Id096c7b5f330d6abd8f5b01c96448a9d8aac8001
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149616
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>
(cherry picked from commit f0287aad967761d46cb6903fc0d985a536408176)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150716
Tested-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>