sw floattable: fix removal of empty master flys on master table removal
The bugdoc has 2 pages, a floating table covering both pages, row 1 on
page 1, row 2 on page 2. Once row 1 was deleted, the text around row 2
moved up to page 1, but row 2 remained on page 2, which is unexpected.

What happened is that we correctly deleted the fly of row 1, even moved
the text of the anchor on page 2 to page 1, but nobody moved the fly of
row 2 to page 1, which is inconsistent.

Fix the problem by invalidating the position of the follow in
SwFlyAtContentFrame::DelEmpty(), since deleting a master fly potentially
creates space for follows, so they can move up.

This helps in the 2-page case, the case with more than 2 pages still
needs work.

(cherry picked from commit 1006fd848ba7c67927472e53df7f3b6f682fadfb)

Change-Id: Ida9cc1d586963867e8fc7ee92d4b3f5dfafd3579
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150009
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>