oox: refactor ThemeExport to use member FSHelperPtr instance
This refactors ThemeExport to use the member variable FSHelperPtr
instead of passign it through functions as a parameter. This also
forces many functions that were in anonymous namespace to become
members of the ThemeExport class. This is needed because in some
instances we will need access to FSHelperPtr instance as well as
access to XmlFilterBase.

Change-Id: I9e4ead9cd87ec2157535106be28596c4282fe167
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150260
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 78cf9c05db01e2177d28eb08b3389e487d84060f)