sw floattable: unconditionally map DOC table pos props to SwFormatFlySplit
Previously we only mapped DOC floating tables to split fly frames when
the heuristics asked for a fly frame, this goes further.

Also add a testcase that shows that the import now works, but the export
needs more work.

In other words, this disables heuristics when the feature flag is on.

(cherry picked from commit 69fe424e426957545a30e1b912c933e1e7693100)

Change-Id: I65018fef7994b283a60b0c38b22c16e2277cb80f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150471
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>