tdf#86630 sw page number wizard: set page style numbering type
This is a squashed commit - including
commit d4cf5814146e375805df98eafba2baf321b9b005
    tdf#86630 sw page number wizard: cleanup numbering type patch
and commit 0b35b7953507491fe1ebdead9ac6ce592bd9b674
    cid#1528191 Uninitialized scalar field

Giving the user the opportunity to change the number type
is rather important. Well, maybe not too critical since I
assume that the default page number style is relatively accurate
for each locale, and at least for English we always just
use numbers for the page style.

However, a wizard is the perfect place to expose such a setting,
otherwise it is buried in page style settings - far from
where anyone would look when inserting a page number.

Change-Id: I9053504fa55d16ebf5424946b480cb044b79988c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151005
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
 Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151899
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>